Remove dynamic children from Course component
